Extract structure into printableData
